+++ src/main/java/com/hazzum/storefront/service/item/ItemService.java
@@ -6,10 +6,10 @@ import com.hazzum.storefront.entity.Item;
 import com.hazzum.storefront.payload.response.CartItem;
 
 public interface ItemService {
-    List<CartItem> showAll(int orderID);
-    Item findById(int itemID);
-    Item addItem(int quantity, int orderID, int productID);
-    Item updateQuantity(int itemID, int quantity);
-    void removeItem(int itemID);
-    void commitOrder(int theId);
+    List<CartItem> showAll(Long orderID);
+    Item findById(Long itemID);
+    Item addItem(int quantity, Long orderID, Long productID);
+    Item updateQuantity(Long itemID, int quantity);
+    void removeItem(Long itemID);
+    void commitOrder(Long theId);
 }
